Refactor ScreenUtilities
