+++ Source/Engine/Platform/Types.h
@@ -4,8 +4,6 @@
 
 #if PLATFORM_WINDOWS
 
-class WindowsClipboard;
-typedef WindowsClipboard Clipboard;
 class Win32CriticalSection;
 typedef Win32CriticalSection CriticalSection;
 class Win32ConditionVariable;
@@ -16,21 +14,23 @@ class WindowsFileSystemWatcher;
 typedef WindowsFileSystemWatcher FileSystemWatcher;
 class Win32File;
 typedef Win32File File;
-class WindowsPlatform;
-typedef WindowsPlatform Platform;
 class Win32Thread;
 typedef Win32Thread Thread;
+class WindowsClipboard;
+typedef WindowsClipboard Clipboard;
+class WindowsPlatform;
+typedef WindowsPlatform Platform;
 class WindowsWindow;
 typedef WindowsWindow Window;
 class Win32Network;
 typedef Win32Network Network;
 class UserBase;
 typedef UserBase User;
+class WindowsScreenUtilities;
+typedef WindowsScreenUtilities ScreenUtilities;
 
 #elif PLATFORM_UWP
 
-class ClipboardBase;
-typedef ClipboardBase Clipboard;
 class Win32CriticalSection;
 typedef Win32CriticalSection CriticalSection;
 class Win32ConditionVariable;
@@ -41,10 +41,12 @@ class FileSystemWatcherBase;
 typedef FileSystemWatcherBase FileSystemWatcher;
 class Win32File;
 typedef Win32File File;
-class UWPPlatform;
-typedef UWPPlatform Platform;
 class Win32Thread;
 typedef Win32Thread Thread;
+class ClipboardBase;
+typedef ClipboardBase Clipboard;
+class UWPPlatform;
+typedef UWPPlatform Platform;
 class UWPWindow;
 typedef UWPWindow Window;
 class Win32Network;
@@ -54,8 +56,6 @@ typedef UserBase User;
 
 #elif PLATFORM_LINUX
 
-class LinuxClipboard;
-typedef LinuxClipboard Clipboard;
 class UnixCriticalSection;
 typedef UnixCriticalSection CriticalSection;
 class UnixConditionVariable;
@@ -66,21 +66,23 @@ class LinuxFileSystemWatcher;
 typedef LinuxFileSystemWatcher FileSystemWatcher;
 class UnixFile;
 typedef UnixFile File;
-class LinuxPlatform;
-typedef LinuxPlatform Platform;
 class LinuxThread;
 typedef LinuxThread Thread;
+class LinuxClipboard;
+typedef LinuxClipboard Clipboard;
+class LinuxPlatform;
+typedef LinuxPlatform Platform;
 class LinuxWindow;
 typedef LinuxWindow Window;
 class UnixNetwork;
 typedef UnixNetwork Network;
 class UserBase;
 typedef UserBase User;
+class LinuxScreenUtilities;
+typedef LinuxScreenUtilities ScreenUtilities;
 
 #elif PLATFORM_PS4
 
-class ClipboardBase;
-typedef ClipboardBase Clipboard;
 class UnixCriticalSection;
 typedef UnixCriticalSection CriticalSection;
 class UnixConditionVariable;
@@ -91,10 +93,12 @@ class FileSystemWatcherBase;
 typedef FileSystemWatcherBase FileSystemWatcher;
 class UnixFile;
 typedef UnixFile File;
-class PS4Platform;
-typedef PS4Platform Platform;
 class PS4Thread;
 typedef PS4Thread Thread;
+class ClipboardBase;
+typedef ClipboardBase Clipboard;
+class PS4Platform;
+typedef PS4Platform Platform;
 class PS4Window;
 typedef PS4Window Window;
 class PS4Network;
@@ -104,8 +108,6 @@ typedef PS4User User;
 
 #elif PLATFORM_PS5
 
-class ClipboardBase;
-typedef ClipboardBase Clipboard;
 class UnixCriticalSection;
 typedef UnixCriticalSection CriticalSection;
 class UnixConditionVariable;
@@ -116,10 +118,12 @@ class FileSystemWatcherBase;
 typedef FileSystemWatcherBase FileSystemWatcher;
 class UnixFile;
 typedef UnixFile File;
-class PS5Platform;
-typedef PS5Platform Platform;
+class ClipboardBase;
+typedef ClipboardBase Clipboard;
 class PS5Thread;
 typedef PS5Thread Thread;
+class PS5Platform;
+typedef PS5Platform Platform;
 class PS5Window;
 typedef PS5Window Window;
 class PS5Network;
@@ -129,8 +133,6 @@ typedef PS5User User;
 
 #elif PLATFORM_XBOX_ONE
 
-class ClipboardBase;
-typedef ClipboardBase Clipboard;
 class Win32CriticalSection;
 typedef Win32CriticalSection CriticalSection;
 class Win32ConditionVariable;
@@ -141,10 +143,12 @@ class FileSystemWatcherBase;
 typedef FileSystemWatcherBase FileSystemWatcher;
 class Win32File;
 typedef Win32File File;
-class XboxOnePlatform;
-typedef XboxOnePlatform Platform;
 class Win32Thread;
 typedef Win32Thread Thread;
+class ClipboardBase;
+typedef ClipboardBase Clipboard;
+class XboxOnePlatform;
+typedef XboxOnePlatform Platform;
 class GDKWindow;
 typedef GDKWindow Window;
 class Win32Network;
@@ -154,8 +158,6 @@ typedef GDKUser User;
 
 #elif PLATFORM_XBOX_SCARLETT
 
-class ClipboardBase;
-typedef ClipboardBase Clipboard;
 class Win32CriticalSection;
 typedef Win32CriticalSection CriticalSection;
 class Win32ConditionVariable;
@@ -166,10 +168,12 @@ class FileSystemWatcherBase;
 typedef FileSystemWatcherBase FileSystemWatcher;
 class Win32File;
 typedef Win32File File;
-class XboxScarlettPlatform;
-typedef XboxScarlettPlatform Platform;
 class Win32Thread;
 typedef Win32Thread Thread;
+class ClipboardBase;
+typedef ClipboardBase Clipboard;
+class XboxScarlettPlatform;
+typedef XboxScarlettPlatform Platform;
 class GDKWindow;
 typedef GDKWindow Window;
 class Win32Network;
@@ -179,8 +183,6 @@ typedef GDKUser User;
 
 #elif PLATFORM_ANDROID
 
-class ClipboardBase;
-typedef ClipboardBase Clipboard;
 class UnixCriticalSection;
 typedef UnixCriticalSection CriticalSection;
 class UnixConditionVariable;
@@ -191,10 +193,12 @@ class FileSystemWatcherBase;
 typedef FileSystemWatcherBase FileSystemWatcher;
 class AndroidFile;
 typedef AndroidFile File;
-class AndroidPlatform;
-typedef AndroidPlatform Platform;
 class AndroidThread;
 typedef AndroidThread Thread;
+class ClipboardBase;
+typedef ClipboardBase Clipboard;
+class AndroidPlatform;
+typedef AndroidPlatform Platform;
 class AndroidWindow;
 typedef AndroidWindow Window;
 class UnixNetwork;
@@ -204,8 +208,6 @@ typedef UserBase User;
 
 #elif PLATFORM_SWITCH
 
-class ClipboardBase;
-typedef ClipboardBase Clipboard;
 class SwitchCriticalSection;
 typedef SwitchCriticalSection CriticalSection;
 class SwitchConditionVariable;
@@ -216,10 +218,12 @@ class FileSystemWatcherBase;
 typedef FileSystemWatcherBase FileSystemWatcher;
 class SwitchFile;
 typedef SwitchFile File;
-class SwitchPlatform;
-typedef SwitchPlatform Platform;
 class SwitchThread;
 typedef SwitchThread Thread;
+class ClipboardBase;
+typedef ClipboardBase Clipboard;
+class SwitchPlatform;
+typedef SwitchPlatform Platform;
 class SwitchWindow;
 typedef SwitchWindow Window;
 class SwitchNetwork;
@@ -229,8 +233,6 @@ typedef SwitchUser User;
 
 #elif PLATFORM_MAC
 
-class MacClipboard;
-typedef MacClipboard Clipboard;
 class UnixCriticalSection;
 typedef UnixCriticalSection CriticalSection;
 class UnixConditionVariable;
@@ -241,21 +243,23 @@ class MacFileSystemWatcher;
 typedef MacFileSystemWatcher FileSystemWatcher;
 class UnixFile;
 typedef UnixFile File;
-class MacPlatform;
-typedef MacPlatform Platform;
 class AppleThread;
 typedef AppleThread Thread;
+class MacClipboard;
+typedef MacClipboard Clipboard;
+class MacPlatform;
+typedef MacPlatform Platform;
 class MacWindow;
 typedef MacWindow Window;
 class UnixNetwork;
 typedef UnixNetwork Network;
 class UserBase;
 typedef UserBase User;
+class MacScreenUtilities;
+typedef MacScreenUtilities ScreenUtilities;
 
 #elif PLATFORM_IOS
 
-class ClipboardBase;
-typedef ClipboardBase Clipboard;
 class UnixCriticalSection;
 typedef UnixCriticalSection CriticalSection;
 class UnixConditionVariable;
@@ -266,10 +270,12 @@ class FileSystemWatcherBase;
 typedef FileSystemWatcherBase FileSystemWatcher;
 class iOSFile;
 typedef iOSFile File;
-class iOSPlatform;
-typedef iOSPlatform Platform;
 class AppleThread;
 typedef AppleThread Thread;
+class ClipboardBase;
+typedef ClipboardBase Clipboard;
+class iOSPlatform;
+typedef iOSPlatform Platform;
 class iOSWindow;
 typedef iOSWindow Window;
 class UnixNetwork;
